Refine code.
The size of got entry and fragment is hard to out of the range of size_t.
So, I use size_t instead of uint64_t.

+++ include/mcld/MC/MCTargetFragment.h
@@ -32,7 +32,7 @@ protected:
 public:
   virtual ~MCTargetFragment() {}
 
-  virtual uint64_t getSize() const = 0;
+  virtual size_t getSize() const = 0;
 
 public:
   static bool classof(const MCFragment *F)
